+++ container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider/enclave_object_attributes_provider.go
@@ -15,10 +15,12 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/module"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
+	"github.com/kurtosis-tech/container-engine-lib/lib/uuid_generator"
 	"github.com/kurtosis-tech/stacktrace"
 )
 
 const (
+	namespacePrefix = "kurtosis-enclave"
 	artifactExpansionObjectTimestampFormat = "2006-01-02T15.04.05.000"
 	filesArtifactExpansionPrefix = "files-artifact-expansion"
 	userServicePrefix = "user-service"
@@ -77,7 +79,14 @@ func GetKubernetesEnclaveObjectAttributesProvider(enclaveId enclave.EnclaveID) K
 }
 
 func (provider *kubernetesEnclaveObjectAttributesProviderImpl) ForEnclaveNamespace(isPartitioningEnabled bool) (KubernetesObjectAttributes, error) {
-	name, err := kubernetes_object_name.CreateNewKubernetesObjectName(provider.enclaveId)
+	namespaceUuid, err := uuid_generator.GenerateUUIDString()
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "Failed to generate UUID string for namespace name for enclave '%v'", provider.enclaveId)
+	}
+	name, err := getCompositeKubernetesObjectName([]string{
+		namespacePrefix,
+		namespaceUuid,
+	})
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred creating a name object from string '%v'", provider.enclaveId)
 	}